@@ -85,9 +85,9 @@ class PostBannerActivity : BaseActivity(), BannerView.Post {
 
     override fun loadingIndicator(isLoading: Boolean) {
         if(isLoading){
-            layout_progress.visibility = View.VISIBLE
+            layoutProgress.visibility = View.VISIBLE
         }else{
-            layout_progress.visibility = View.GONE
+            layoutProgress.visibility = View.GONE
         }
     }
 
