@@ -71,7 +71,7 @@ public class ListBasketActivity extends BaseActivity implements ProductAdapter.O
         etSearch = findViewById(R.id.etSearchView);
         recyclerView = findViewById(R.id.recyclerView);
         progress_bar = findViewById(R.id.progress_bar);
-        ll_no_product = findViewById(R.id.layout_empty);
+        ll_no_product = findViewById(R.id.layoutEmpty);
         swipe_refresh =findViewById(R.id.swipeRefresh);
         nestedScrollView = findViewById(R.id.nestedScrollView);
         rl_load_more = findViewById(R.id.rlLoadMore);
